requirements.txt
asgiref==3.7.2
Babel==2.12.1
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.5
django-crispy-forms==2.0
django-environ==0.11.2
django-heroku==0.3.1
django-import-export==3.3.1
django-phonenumber-field==7.1.0
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
phonenumberslite==8.13.21
Pillow==10.0.1
psycopg2==2.9.7
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
